requirements.txt
torchtext
transformers
pillow
transformers
matplotlib
torchdata
tqdm